+++ kitty/glfw.c
@@ -135,6 +135,7 @@ set_layer_shell_config_for(OSWindow *w, GLFWLayerShellConfig *lsc) {
         lsc->related.background_opacity = w->background_opacity;
         lsc->related.background_blur = OPT(background_blur);
         lsc->related.color_space = OPT(macos_colorspace);
+        w->hide_on_focus_loss = lsc->hide_on_focus_loss;
     }
     return glfwSetLayerShellConfig(w->handle, lsc);
 }
@@ -562,10 +563,17 @@ set_os_window_visibility(OSWindow *w, int set_visible) {
 }
 
 static void
+update_os_window_visibility_based_on_focus(id_type timer_id UNUSED, void*d) {
+    OSWindow * osw = os_window_for_id((uintptr_t)d);
+    if (osw && osw->hide_on_focus_loss && !osw->is_focused) set_os_window_visibility(osw, 0);
+}
+
+static void
 window_focus_callback(GLFWwindow *w, int focused) {
     if (!set_callback_window(w)) return;
 #define osw global_state.callback_os_window
     debug_input("\x1b[35mon_focus_change\x1b[m: window id: 0x%llu focused: %d\n", osw->id, focused);
+    bool focus_changed = osw->is_focused != focused;
     osw->is_focused = focused ? true : false;
     monotonic_t now = monotonic();
     id_type wid = osw->id;
@@ -591,8 +599,8 @@ window_focus_callback(GLFWwindow *w, int focused) {
         }
     }
     request_tick_callback();
-    if (osw && osw->hide_on_focus_lost && osw->handle) {
-        if (glfwGetWindowAttrib(osw->handle, GLFW_VISIBLE)) set_os_window_visibility(osw, 0);
+    if (osw && osw->handle && !focused && focus_changed && osw->hide_on_focus_loss && glfwGetWindowAttrib(osw->handle, GLFW_VISIBLE)) {
+        add_main_loop_timer(0, false, update_os_window_visibility_based_on_focus, (void*)(uintptr_t)osw->id, NULL);
     }
     osw = NULL;
 #undef osw
@@ -1216,6 +1224,7 @@ layer_shell_config_from_python(PyObject *p, GLFWLayerShellConfig *ans) {
     A(requested_right_margin, PyLong_Check, PyLong_AsLong);
     A(requested_exclusive_zone, PyLong_Check, PyLong_AsLong);
     A(override_exclusive_zone, PyBool_Check, PyLong_AsLong);
+    A(hide_on_focus_loss, PyBool_Check, PyLong_AsLong);
 #undef A
 #define A(attr) { \
     RAII_PyObject(attr, PyObject_GetAttrString(p, #attr)); if (attr == NULL) return false; \
@@ -1255,17 +1264,11 @@ create_os_window(PyObject UNUSED *self, PyObject *args, PyObject *kw) {
         window_state = (int) PyLong_AsLong(optional_window_state);
     }
     if (layer_shell_config && layer_shell_config != Py_None ) {
-#ifdef __APPLE__
-        lsc = &lsc_stack;
-#else
-        if (global_state.is_wayland) {
-            if (!glfwWaylandIsLayerShellSupported()) {
-                PyErr_SetString(PyExc_RuntimeError, "The Wayland compositor does not support the layer shell protocol.");
-                return NULL;
-            }
-            lsc = &lsc_stack;
+        if (!glfwIsLayerShellSupported()) {
+            PyErr_SetString(PyExc_RuntimeError, "The window manager/compositor does not support the primitives needed to make panels.");
+            return NULL;
         }
-#endif
+        lsc = &lsc_stack;
     } else {
         if (optional_x && optional_x != Py_None) { if (!PyLong_Check(optional_x)) { PyErr_SetString(PyExc_TypeError, "x must be an int"); return NULL;} x = (int)PyLong_AsLong(optional_x); }
         if (optional_y && optional_y != Py_None) { if (!PyLong_Check(optional_y)) { PyErr_SetString(PyExc_TypeError, "y must be an int"); return NULL;} y = (int)PyLong_AsLong(optional_y); }
@@ -1413,7 +1416,10 @@ create_os_window(PyObject UNUSED *self, PyObject *args, PyObject *kw) {
     OSWindow *w = add_os_window();
     w->handle = glfw_window;
     w->disallow_title_changes = disallow_override_title;
-    w->is_layer_shell = lsc != NULL;
+    if (lsc != NULL) {
+        w->is_layer_shell = true;
+        w->hide_on_focus_loss = lsc->hide_on_focus_loss;
+    }
     update_os_window_references();
     if (!w->is_layer_shell || (global_state.is_apple && w->is_layer_shell && lsc->focus_policy == GLFW_FOCUS_EXCLUSIVE)) {
         for (size_t i = 0; i < global_state.num_os_windows; i++) {
@@ -2499,28 +2505,8 @@ get_clipboard_mime(PyObject *self UNUSED, PyObject *args) {
 }
 
 static PyObject*
-make_x11_window_a_dock_window(PyObject *self UNUSED, PyObject *args UNUSED) {
-    int x11_window_id;
-    PyObject *dims;
-    if (!PyArg_ParseTuple(args, "iO!", &x11_window_id, &PyTuple_Type, &dims)) return NULL;
-    if (PyTuple_GET_SIZE(dims) != 12 ) { PyErr_SetString(PyExc_TypeError, "dimensions must be a tuple of length 12"); return NULL; }
-    if (!glfwSetX11WindowAsDock) { PyErr_SetString(PyExc_RuntimeError, "Failed to load glfwGetX11Window"); return NULL; }
-    uint32_t dimensions[12];
-    for (Py_ssize_t i = 0; i < 12; i++) dimensions[i] = PyLong_AsUnsignedLong(PyTuple_GET_ITEM(dims, i));
-    if (PyErr_Occurred()) return NULL;
-    glfwSetX11WindowAsDock(x11_window_id);
-    glfwSetX11WindowStrut(x11_window_id, dimensions);
-    Py_RETURN_NONE;
-}
-
-static PyObject*
 is_layer_shell_supported(PyObject *self UNUSED, PyObject *args UNUSED) {
-#ifdef __APPLE__
-    Py_RETURN_FALSE;
-#else
-    if (!global_state.is_wayland) Py_RETURN_FALSE;
-    return Py_NewRef(glfwWaylandIsLayerShellSupported() ? Py_True : Py_False);
-#endif
+    return Py_NewRef(glfwIsLayerShellSupported() ? Py_True : Py_False);
 }
 
 static PyObject*
@@ -2565,23 +2551,12 @@ set_layer_shell_config(PyObject *self UNUSED, PyObject *args) {
     return Py_NewRef(set_layer_shell_config_for(window, &lsc) ? Py_True : Py_False);
 }
 
-static PyObject*
-set_os_window_hide_on_focus_lost(PyObject *self UNUSED, PyObject *args) {
-    unsigned long long wid; int val = 1;
-    if (!PyArg_ParseTuple(args, "K|p", &wid, &val)) return NULL;
-    OSWindow *window = os_window_for_id(wid);
-    if (!window) Py_RETURN_FALSE;
-    window->hide_on_focus_lost = val;
-    Py_RETURN_TRUE;
-}
-
 
 // Boilerplate {{{
 
 static PyMethodDef module_methods[] = {
     METHODB(set_custom_cursor, METH_VARARGS),
     METHODB(is_css_pointer_name_valid, METH_O),
-    METHODB(set_os_window_hide_on_focus_lost, METH_O),
     METHODB(toggle_os_window_visibility, METH_VARARGS),
     METHODB(layer_shell_config_for_os_window, METH_O),
     METHODB(set_layer_shell_config, METH_VARARGS),
@@ -2603,7 +2578,6 @@ static PyMethodDef module_methods[] = {
     METHODB(get_click_interval, METH_NOARGS),
     METHODB(is_layer_shell_supported, METH_NOARGS),
     METHODB(x11_window_id, METH_O),
-    METHODB(make_x11_window_a_dock_window, METH_VARARGS),
     METHODB(strip_csi, METH_O),
 #ifndef __APPLE__
     METHODB(dbus_close_notification, METH_VARARGS),